+++ plugins/inputs/proxmox/structs.go
@@ -2,34 +2,12 @@ package proxmox
 
 import (
 	"encoding/json"
-	"net/http"
-	"net/url"
-
-	"github.com/influxdata/telegraf"
-	"github.com/influxdata/telegraf/config"
-	"github.com/influxdata/telegraf/plugins/common/tls"
 )
 
 var (
 	qemu resourceType = "qemu"
 	lxc  resourceType = "lxc"
 )
-
-type Proxmox struct {
-	BaseURL               string          `toml:"base_url"`
-	APIToken              string          `toml:"api_token"`
-	ResponseTimeout       config.Duration `toml:"response_timeout"`
-	NodeName              string          `toml:"node_name"`
-	AdditionalVmstatsTags []string        `toml:"additional_vmstats_tags"`
-	tls.ClientConfig
-
-	Log telegraf.Logger `toml:"-"`
-
-	httpClient *http.Client
-
-	nodeSearchDomain string
-	requestFunction  func(px *Proxmox, apiUrl string, method string, data url.Values) ([]byte, error)
-}
 
 type resourceType string
